requirements.txt
blinker==1.9.0
click==8.2.1
colorama==0.4.6
distlib==0.4.0
filelock==3.18.0
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
greenlet==3.2.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
numpy==2.3.2
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
Pygments==2.19.2
qrcode==8.2
setuptools==80.9.0
SQLAlchemy==2.0.41
svgwrite==1.4.3
typing_extensions==4.14.1
virtualenv==20.32.0
Werkzeug==3.1.3
wheel==0.45.1
